декомпозировал файл корневой саги
