+++ frontend/src/config/configParams.js
@@ -1,0 +1,2 @@
+export const MAX_ATTEMPTS = parseInt(import.meta.env.VITE_REQUEST_MAX_ATTEMPTS) || 3;
+export const RETRY_DELAY_MS = parseInt(import.meta.env.VITE_RETRY_DELAY_MS) || 3000;
